refactor(admin): extract UserManagement and ReportsModeration
Move user list (incl. servers modal) and reports sections into
self-contained components. Cross-module refreshes now go through
admin:users-changed / admin:activity-changed window events instead of
reaching into the parent.
